requirements.txt
Flask>=2.2.5
Flask-SQLAlchemy>=3.1.1
Flask-Migrate
requests
Flask-WTF>=1.2.1
razorpay>=1.4.2
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
